Revert "moved driving commands into their own folder"
This reverts commit 7bfd4211a1f930a16f502c567ebe50c63861ca78.
