+++ src/main/java/frc/robot/commands/LockDrivetrain.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.drive_commands;
+package frc.robot.commands;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
