+++ src/main/java/frc/robot/commands/SwerveDriveCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.drive_commands;
+package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.drivetrain.SwerveDrivetrain;
